move props to component file
