--- components/types/HeroButtonProps.ts
@@ -1,6 +1,0 @@
-export type HeroButtonProps = {
-  children: React.ReactNode
-  href: string
-  icon: any
-  main?: boolean
-}
