--- components/types/NewsListProps.ts
@@ -1,6 +1,0 @@
-import { News } from "@/functions/news"
-
-export type NewsListProps = {
-  news: News[]
-  href: string
-}
